refactor document loaders to utils
